requirements.txt
# MT5 Trading Bot for FTMO
# Core
MetaTrader5>=5.0.45; sys_platform == 'win32'
pandas>=2.1.0
numpy>=1.25.0

# Technical Analysis — SMC implemented in src/strategy/smc_engine.py (pandas).
# smartmoneyconcepts pulls numba/llvmlite; optional on Windows if you switch engines.
# pandas-ta>=0.3.14b  # optional indicators

# Async
asyncio-mqtt>=0.16.0
aiohttp>=3.9.0
aiosqlite>=0.19.0

# Telegram
python-telegram-bot>=21.0

# Config & Environment
pyyaml>=6.0
python-dotenv>=1.0.0

# Scheduling
APScheduler>=3.10.0

# Logging
loguru>=0.7.0

# Data validation
pydantic>=2.5.0
